refactor: inline updateUser into settings page, remove useUser hook
